requirements.txt
fastapi>=0.85.0
uvicorn>=0.18.3
folium>=
pymongo[srv]>=4.2.0
python-dotenv>=0.21.0
mypy>=0.982
pytest>=7.1.3
requests>=2.28.1
Jinja2>=3.1.2
autopep8>=1.7.0
pandas>=1.5.0
imgurpython>=1.1.7
selenium>=4.6.0